+++ client/e2e/tsconfig.json
