+++ client/src/app/display-resource/display-resource.component.scss
@@ -1,0 +1,183 @@
+.select-region {
+    border: solid 1px #ccc;
+    margin: 10%;
+    height: 60%;
+    border-radius: 10px;
+    display: block;
+  }
+
+.select-region-highlight {
+  border: solid 3px #F29111;
+  border-radius: 10px;
+  margin: 10%;
+  margin-bottom: 4%;
+  height: 60%;
+  display: block;
+}
+
+.select-region:hover {
+    background-color: gray;
+  }
+
+.select-region .centered {
+  position: relative;
+  top: 50%;
+  text-align: center;
+}
+h3 {
+  font-family: 'Roboto', sans-serif;
+  font-size: 12px;
+  color: #939699;
+  font-weight: 400;
+  
+  
+}
+.decorate-region-text {
+  border: solid 3px #F29111;
+  border-radius: 24px;
+  width: 80px;
+  height: 24px;
+  padding-left: 14px;
+  margin-top:-16px;
+  margin-left:41px;
+  background:white;
+  font-family: 'Roboto', sans-serif;
+  font-size: 12px;
+  color: #F29111;
+  font-weight: bold;
+}
+
+.decorate-ad-text {
+  border: solid 3px #00758F;
+  border-radius: 24px;
+  width: 80px;
+  height: 24px;
+  padding-left: 14px;
+  margin-top:-16px;
+  margin-left:41px;
+  background:white;
+  font-family: 'Roboto', sans-serif;
+  font-size: 12px;
+  color: #00758F;
+  font-weight: bold;
+  
+}
+
+
+
+.select-ad {
+  border: solid 1px #ccc;
+  border-radius: 10px;
+  margin: 40px;
+  height: 80%;
+  display: block;
+}
+
+.select-ad .centered {
+  position: relative;
+  top: 50%;
+  text-align: center;
+}
+
+.select-ad:hover {
+  background-color: gray;
+}
+.select-ad-highlight {
+  border: solid 3px #00758F;
+  border-radius: 10px;
+  margin: 40px;
+  height: 80%;
+  display: block;
+}
+
+.select-tier {
+  border: solid 1px #ccc;
+  margin-left: 40px;
+  margin-right: 40px;
+  margin-top: 38px;
+  height: 72%;
+  border-radius: 10px;
+  display: block;
+}
+
+.select-tier .centered {
+  position: relative;
+  top: 50%;
+  text-align: center;
+}
+
+.select-tier:hover {
+  background-color: gray;
+}
+.select-tier-highlight {
+  border: solid 3px #3A913F;
+  margin-left: 40px;
+  margin-right: 40px;
+  margin-top: 38px;
+  height: 72%;
+  border-radius: 10px;
+  display: block;
+}
+
+.decorate-tier-text {
+  border: solid 3px #3A913F;
+  border-radius: 24px;
+  width: 80px;
+  height: 24px;
+  padding-left: 14px;
+  margin-top:-16px;
+  margin-left:41px;
+  background:white;
+  font-family: 'Roboto', sans-serif;
+  font-size: 12px;
+  color: #3A913F;
+  font-weight: bold;
+  
+}
+
+.example-box {
+  // background-image: url('http://source.unsplash.com/random/100x100');
+  width: 100px;
+  height: 100px;
+  border: solid 1px #ccc;
+  color: #F29111;
+  cursor: move;
+  display: inline-flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  background: #fff;
+  border-radius: 4px;
+  margin-top: 25px;
+  margin-left: 25px;
+  margin-right: 25px;
+  margin-bottom: 25px;
+  position: relative;
+  z-index: 1;
+  box-sizing: border-box;
+  font-family: 'Roboto', sans-serif;
+  font-weight: bold;
+  font-size: 14px;
+  background-image: url('../../assets/region.png');
+  background-repeat: no-repeat;
+  padding: 10px;
+  transition: box-shadow 200ms cubic-bezier(0, 0, 0.2, 1);
+  box-shadow: 0 3px 1px -2px rgba(0, 0, 0, 0.2),
+              0 2px 2px 0 rgba(0, 0, 0, 0.14),
+              0 1px 5px 0 rgba(0, 0, 0, 0.12);
+}
+
+button {
+  z-index: 100;
+  margin-left: 40%;
+  margin-bottom: 33%;
+  background-color: #00758F;
+  color: #FFFFFF;
+  border: none;
+  border-radius: 15px;
+  height: 32px;
+  width: 80px;
+  font-family: 'Roboto', sans-serif;
+  font-size: 14px;
+  font-weight: bold;
+}
