+++ client/src/app/divider/divider.component.html
@@ -1,0 +1,49 @@
+<div class=title>
+  Customer Experience
+</div>
+<hr>
+<div cdkDropListGroup>
+  <div id="col-1">
+    <div id="carouselExampleIndicators" class="carousel slide" data-ride="carousel" data-interval="false">
+      <ol class="carousel-indicators">
+        <li data-target="#carouselExampleIndicators" id='region' class="active" data-slide="next"></li>
+        <li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
+        <li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
+        <li data-target="#carouselExampleIndicators" data-slide-to="3"></li>
+      </ol>
+      <div class="carousel-inner">
+        <div class="carousel-item active">
+          <div class="display-text">Region</div>
+          <div class="display-smaller-text">Select a region</div>
+          <app-resource-selector [items]="items" [isTier]="false" ></app-resource-selector>
+        </div>
+        <div class="carousel-item">
+            <div class="display-text">Region</div>
+            <div class="display-smaller-text">Select a region</div>
+            <app-resource-selector [items]="ads" [isTier]="false"></app-resource-selector>
+        </div>
+        <div class="carousel-item">
+            <div class="display-text">Region</div>
+            <div class="display-smaller-text">Select a region</div>
+            <app-resource-selector [items]="tiers" [isTier]="false"></app-resource-selector>
+        </div>
+        <div class="carousel-item">
+          <div class="display-text">Region</div>
+          <div class="display-smaller-text">Select a region</div>
+          <app-resource-selector [items]="platforms"></app-resource-selector>
+        </div>
+        
+        
+      </div>
+      
+    </div>    
+  </div>
+
+  <div id="col-2">
+    <app-display-resource [region]="region" [ad]="ad" [tier]="tier" [platform]="platform" (emittedRegion)='getRegion($event)' (emittedAd)="getAd($event)" (emittedTier)="getTier($event)" (emittedPlatform)="getPlatform($event)" ></app-display-resource>
+  </div>
+  
+
+  
+</div>
+
