+++ client/src/app/divider/divider.component.scss
@@ -1,0 +1,78 @@
+.title {
+  height: 60px;
+  width: 27%;
+  left: 0;
+  background-color: #EEEFF0;
+  color: #4E5052;
+  font-family: 'Roboto', sans-serif;
+  font-size:  17px;
+  font-weight: bold;
+  padding-left: 28px;
+  padding-top: 14px;
+}
+#col-1 {
+  position: fixed;
+  width: 27%;
+  left: 0;
+  height: 100%;
+  background-color: #EEEFF0;
+  
+}
+
+#col-2 {
+  position: fixed;
+  width: 73%;
+  left: 27%;
+  height: 100%;
+  background-color: white;
+  
+}
+
+
+hr {
+  margin: 0px;
+  height: 1px;
+  width: 27%;
+  
+
+}
+.carousel-indicators {
+  bottom: unset;
+}
+
+.carousel-indicators li {
+  display: inline-block;
+  background-color: #4E5052;
+  width: 10px;
+  height: 10px;
+  margin-top: 40px;
+  margin-bottom: 40px;
+  margin-left: 10px;
+  margin-right: 30px;
+  border: 1px solid #fff;
+  border-radius: 10px;
+  cursor: pointer;
+}
+
+.display-text {
+  margin-top:80px;
+  margin-left:40px;
+  margin-right: 40px;
+  padding-left: 100px;
+  font-family: 'Roboto', sans-serif;
+  font-size: 20px;
+  font-weight: bold;
+  color: #F29111;
+}
+
+.display-smaller-text {
+  font-family: 'Roboto', sans-serif;
+  font-size: 12px;
+  color: #939699;
+  font-weight: 400;
+  margin-left:40px;
+  padding-left: 40px;
+}
+
+
+
