+++ client/src/app/resource-selector/resource-selector.component.spec.ts
