+++ client/src/environments/environment.prod.ts
