--- src/app/resource-selector/resource-selector.component.scss
@@ -1,67 +1,0 @@
-.example-list {
-    
-    margin-top: 10px;
-    height: 100%;
-    display: block;
-  }
-
-.example-box {
-    // background-image: url('http://source.unsplash.com/random/100x100');
-    width: 140px;
-    height: 140px;
-    border: solid 1px #ccc;
-    color: #F29111;
-    cursor: move;
-    display: inline-flex;
-    justify-content: center;
-    align-items: flex-end;
-    text-align: center;
-    background: #fff;
-    border-radius: 4px;
-    margin-top: 20px;
-    margin-left: 25px;
-    margin-right: 25px;
-    margin-bottom: 20px;
-    position: relative;
-    z-index: 1;
-    box-sizing: border-box;
-    padding: 10px;
-    font-family: 'Roboto', sans-serif;
-    background-image: url('../../assets/region.png');
-    background-repeat: no-repeat;
-    background-position-x: 25px;
-    background-size: 50px;
-    background-position-y: 20px;
-    font-size: 14px;
-    font-weight: bold;
-    transition: box-shadow 200ms cubic-bezier(0, 0, 0.2, 1);
-    box-shadow: 0 3px 1px -2px rgba(0, 0, 0, 0.2),
-                0 2px 2px 0 rgba(0, 0, 0, 0.14),
-                0 1px 5px 0 rgba(0, 0, 0, 0.12);
-  }
-  .cdk-drag-preview {
-    box-sizing: border-box;
-    border-radius: 4px;
-    box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2),
-                0 8px 10px 1px rgba(0, 0, 0, 0.14),
-                0 3px 14px 2px rgba(0, 0, 0, 0.12);
-  }
-  
-  .cdk-drag-placeholder {
-    opacity: 0;
-  }
-  
-  .cdk-drag-animating {
-    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
-  }
-  
-  // .example-box:last-child {
-  //   border: none;
-  // }
-  
-  .example-list.cdk-drop-list-dragging .example-box:not(.cdk-drag-placeholder) {
-    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
-  }
-
-  
-
